Rebased and moved the checker to a filter
