+++ sys/parsoid.js
@@ -184,25 +184,6 @@ PSP._dependenciesUpdate = function(hyper, req) {
  * @private
  */
 PSP._wrapInAccessCheck = function(hyper, req, promise) {
-    // TODO: Enable new checking style when the restrictions table is filled
-    // TODO: move to a request filter
-    /*var rp = req.params;
-    var checkURIParts = [rp.domain, 'sys', 'page_revisions', 'restriction', rp.title];
-    if (rp.revision) {
-        checkURIParts.push(rp.revision);
-    }
-
-    return P.join(
-        promise,
-        hyper.get({ uri: new URI(checkURIParts) })
-    )
-    .spread(function(content, restriction) {
-        if (restriction.body && restriction.body.items && restriction.body.items.length) {
-            var revInfo = mwUtil.parseETag(content.headers.etag);
-            mwUtil.applyAccessChecks(restriction.body.items[0], revInfo.rev);
-        }
-        return content;
-    });*/
     return P.props({
         content: promise,
         revisionInfo: this.getRevisionInfo(hyper, req)
